make it all right
git-svn-id: https://svn.strategoxt.org/repos/StrategoXT/sugarj/trunk@22966 ba653a0a-7fb7-0310-a048-973b18a224e4

+++ compiler/org/sugarj/driver/FileCommands.java
@@ -36,7 +36,8 @@ public class FileCommands {
   }
 
   public static void delete(String file) throws IOException {
-    new File(file).delete();
+    if (file != null)
+      new File(file).delete();
   }
 
   public static void copyFile(String from, String to) throws IOException {
